Change llms txt files to folllow semi-standards more closely
